Move useful method to a util module.
